Improve footer layout
The page footer is now the width of the page.

Right-aligning the text looks much neater, but is a bit strange when the
text within the main section of the page is left-aligned. This might be
something to revisit, however large blocks of justified text is
generally harder to read.